WebIDL/tests/submissions/heycam/unrestricted-double-002.html

<!DOCTYPE html>
<title>unrestricted double</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.15 unrestricted double
//
// Store and retrieve a JS Number as an IDL double value through
// a typed array.

var values = [
  undefined,           NaN,                "undefined",
  null,                0,                  "null",
  false,               0,                  "false",
  true,                1,                  "true",
  0,                   0,                  "0",
  1,                   1,                  "1",
  -1,                  -1,                 "-1",
  0.5,                 0.5,                "0.5",
  Infinity,            Infinity,           "Infinity",
  -Infinity,           -Infinity,          "-Infinity",
  NaN,                 NaN,                "NaN",
  -0,                  -0,                 "-0",
  "",                  0,                  '""',
  "1",                 1,                  '"1"',
  "hello",             NaN,                '"hello"',
  { },                 NaN,                "{ }",
  [],                  0,                  "[]",
  /abc/,               NaN,                "/abc/",
  document,            NaN,                "document",
  -2147483648,         -2147483648,        "-2147483648",
  -2147483649,         -2147483649,        "-2147483649",
  4294967295,          4294967295,         "4294967295",
  4294967296,          4294967296,         "4294967296",
  4294967297,          4294967297,         "4294967297",
  -0.5,                -0.5,               "-0.5",
  -1.5,                -1.5,               "-1.5",
  1.5,                 1.5,                "1.5",
  1.1,                 1.1,                "1.1",
  1e309,               Infinity,           "1e309",
  1.0000001788139343,  1.0000001788139343, "1.0000001788139343"
];

// feature: Float64Array (indexed property setter) (of type unrestricted double)

var a = new Float64Array(1);

for (var i = 0; i < values.length; i += 3) {
  var value = values[i];
  var expected = values[i + 1];
  var desc = values[i + 2];
  test(function() {
    a[0] = value;
    assert_equals(a[0], expected);
  }, "passing " + desc + " through a Float64Array indexed property setter coerces the value to an IDL unrestricted double correctly");
}

for (var i = 0; i < values.length; i += 3) {
  var value = values[i];
  var expected = values[i + 1];
  var desc = values[i + 2];
  test(function() {
    var objectValue = { valueOf: function() { return value; }, toString: function() { return { } } };
    if (value === null || typeof value != "object") {
      a[0] = objectValue;
      assert_equals(a[0], expected);
    } else {
      // ToPrimitive() will throw a TypeError if valueOf and toString both return an object.
      assert_throws(new TypeError(), function() { a[0] = objectValue; });
    }
  }, "passing { valueOf: function() { return " + desc + " } } through a Float64Array indexed property setter coerces the value to an IDL unrestricted double correctly");
}

for (var i = 0; i < values.length; i += 3) {
  var value = values[i];
  var expected = values[i + 1];
  var desc = values[i + 2];
  test(function() {
    var objectValue = { valueOf: function() { return { } }, toString: function() { return value; } };
    if (value === null || typeof value != "object") {
      a[0] = objectValue;
      assert_equals(a[0], expected);
    } else {
      // ToPrimitive() will throw a TypeError if valueOf and toString both return an object.
      assert_throws(new TypeError(), function() { a[0] = objectValue; });
    }
  }, "passing { toString: function() { return " + desc + " } } through a Float64Array indexed property setter coerces the value to an IDL unrestricted double correctly");
}

test(function() {
  assert_throws({ name: "SomeError" }, function() { a[0] = { toString: function() { throw { name: "SomeError" }; } }; });
}, "passing { toString: function() { throw ... } } to the Float64Array indexed property setter will propagate the exception");
</script>
